fix(SPEK-527): align draft and basket proxy actions
