+++ src/renderer/features/operations/OperationsConfirm/AddPureProxied/ui/Confirmation.tsx
@@ -85,7 +85,7 @@ export const Confirmation = ({ id = 0, secondaryActionButton, hideSignButton, on
         )}
 
         <div className="flex gap-4">
-          {secondaryActionButton}
+          {!hideSignButton && secondaryActionButton}
 
           {!hideSignButton && (
             <SignButton
