+++ src/renderer/features/proxy-remove/ui/RemoveProxy.tsx
@@ -7,9 +7,8 @@ import { Button } from '@/shared/ui';
 import { Modal } from '@/shared/ui-kit';
 import { basketUtils } from '@/entities/basket';
 import { OperationTitle } from '@/entities/chain';
-import { OperationResult, transactionService } from '@/entities/transaction';
+import { OperationResult } from '@/entities/transaction';
 import { walletModel } from '@/entities/wallet';
-import { InitiateDraftButton } from '@/features/drafts';
 import { OperationSign, OperationSubmit } from '@/features/operations';
 import { RemoveProxyConfirmation as Confirmation } from '@/features/operations/OperationsConfirm/RemoveProxy';
 import { removeProxyUtils } from '../lib/remove-proxy-utils';
@@ -32,12 +31,8 @@ export const RemoveProxy = ({ wallet }: Props) => {
   const chainId = useUnit(removeProxyModel.$proxyAccount.map((proxy) => proxy?.chainId ?? null));
 
   const isPureProxiedNeedToBeKilled = useUnit(removeProxyModel.$isPureProxiedNeedToBeKilled);
-  const coreTx = useUnit(removeProxyModel.$coreTx);
-  const api = useUnit(removeProxyModel.$api);
   const signatoryAccount = useUnit(removeProxyModel.form.fields.signatory.$value);
   const wallets = useUnit(walletModel.$wallets);
-
-  const draftCallData = transactionService.getCallDataHex(coreTx, api);
 
   const signatoryWallet = signatoryAccount ? (wallets.find((w) => w.id === signatoryAccount.walletId) ?? null) : null;
   const isBasketAvailable = signatoryWallet ? basketUtils.isBasketAvailable(signatoryWallet) : false;
@@ -99,19 +94,11 @@ export const RemoveProxy = ({ wallet }: Props) => {
         {removeProxyUtils.isConfirmStep(step) && (
           <Confirmation
             secondaryActionButton={
-              <>
-                {isBasketAvailable && (
-                  <Button pallet="secondary" onClick={() => removeProxyModel.txSaved()}>
-                    {t('operation.addToBasket')}
-                  </Button>
-                )}
-                <InitiateDraftButton
-                  callData={draftCallData}
-                  chainId={chainId}
-                  source="remove-proxy"
-                  onDraftCreated={closeModal}
-                />
-              </>
+              isBasketAvailable ? (
+                <Button pallet="secondary" onClick={() => removeProxyModel.txSaved()}>
+                  {t('operation.addToBasket')}
+                </Button>
+              ) : undefined
             }
             onGoBack={removeProxyModel.wentBackFromConfirm}
           />
